refactor Loader into general Elements module
