+++ src/App/Playlist/index.js
@@ -2,9 +2,9 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import { SpotifyContext } from 'Contexts/index'
 import PageTitle from './PageTitle'
-import Loader from '../Loader'
+import Loader from 'Elements/Loader'
 import LoadedPlaylist from './LoadedPlaylist'
-import getLoaderMessage from '../Loader/sillyExcuses'
+import getLoaderMessage from 'Elements/Loader/sillyExcuses'
 // import MediaPlayer from './MediaPlayer'
 import './Playlist.scss'
